+++ server/src/entities/faction.entity.ts
@@ -6,7 +6,11 @@ import {
   JoinTable,
 } from 'typeorm';
 import { Character } from './character.entity';
-import { ApiProperty, ApiPropertyOptional, ApiHideProperty } from '@nestjs/swagger';
+import {
+  ApiProperty,
+  ApiPropertyOptional,
+  ApiHideProperty,
+} from '@nestjs/swagger';
 
 @Entity()
 export class Faction {