+++ server/src/entities/profile-image.entity.ts
@@ -11,7 +11,11 @@ import {
 } from 'typeorm';
 import { Character } from './character.entity';
 import { User } from './user.entity';
-import { ApiProperty, ApiPropertyOptional, ApiHideProperty } from '@nestjs/swagger';
+import {
+  ApiProperty,
+  ApiPropertyOptional,
+  ApiHideProperty,
+} from '@nestjs/swagger';
 
 @Entity()
 @Index(['isActive'])