+++ server/src/modules/characters/characters.module.ts
@@ -3,10 +3,11 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { CharactersService } from './characters.service';
 import { CharactersController } from './characters.controller';
 import { Character } from '../../entities/character.entity';
+import { Gamble } from '../../entities/gamble.entity';
 import { ServicesModule } from '../../services/services.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Character]), ServicesModule],
+  imports: [TypeOrmModule.forFeature([Character, Gamble]), ServicesModule],
   providers: [CharactersService],
   controllers: [CharactersController],
 })